requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiosqlite==0.19.0
numpy==1.26.2
